RAG AI Document Reader/requirements.txt
streamlit>=1.37.0
openai>=1.0.0
faiss-cpu
pdfplumber>=0.10.2
langchain>=0.2.0
numpy>=1.26.0
